+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 - Support short flags of `--release` (`-r`) and `--features` (`-F`). ([#165](https://github.com/taiki-e/cargo-llvm-cov/pull/165))
 
+- Support [custom profiles](https://doc.rust-lang.org/cargo/reference/profiles.html#custom-profiles). ([#166](https://github.com/taiki-e/cargo-llvm-cov/pull/166))
+
 ## [0.3.3] - 2022-05-06
 
 - Fix an issue where codes in the target directory are not being properly excluded from reports when using `show-env` subcommand. ([#156](https://github.com/taiki-e/cargo-llvm-cov/pull/156))